requirements.txt
apache-airflow==2.10.2
psycopg2-binary==2.9.9
selenium==4.25.0
boto3==1.35.35
botocore==1.35.35
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
apache-airflow-providers-mongo==3.4.0